+++ instrumentation/core/src/androidTest/java/de/mannodermaus/junit5/KotlinInstrumentationTests.kt
@@ -6,7 +6,12 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import de.mannodermaus.junit5.condition.EnabledOnManufacturer
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.DynamicTest.dynamicTest
+import org.junit.jupiter.api.RepeatedTest
+import org.junit.jupiter.api.RepetitionInfo
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestFactory
 import org.junit.jupiter.api.extension.RegisterExtension
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
@@ -25,6 +30,11 @@ class KotlinInstrumentationTests {
     onView(withText("New Text")).check(matches(isDisplayed()))
   }
 
+  @DisplayName("cool display name")
+  @Test
+  fun testWithDisplayName() {
+  }
+
   @Test
   fun testUsingMethodParameter(scenario: ActivityScenario<TestActivity>) {
     onView(withText("TestActivity")).check(matches(isDisplayed()))
@@ -37,6 +47,17 @@ class KotlinInstrumentationTests {
   fun kotlinTestWithParameters(value: Int) {
 
   }
+
+  @RepeatedTest(3)
+  fun kotlinRepeatedTest(info: RepetitionInfo) {
+
+  }
+
+  @TestFactory
+  fun kotlinTestFactory() = listOf(
+    dynamicTest("Dynamic 1") {},
+    dynamicTest("Dynamic 2") {}
+  )
 
   @EnabledOnManufacturer(["Samsung"])
   @Test