+++ instrumentation/runner/src/main/java/de/mannodermaus/junit5/AndroidJUnit5RunnerParams.kt
@@ -2,6 +2,7 @@ package de.mannodermaus.junit5
 
 import org.junit.platform.engine.DiscoverySelector
 import org.junit.platform.engine.Filter
+import org.junit.platform.engine.discovery.MethodSelector
 import org.junit.platform.launcher.LauncherDiscoveryRequest
 import org.junit.platform.launcher.core.LauncherDiscoveryRequestBuilder
 
@@ -13,10 +14,14 @@ data class AndroidJUnit5RunnerParams(
     private val configurationParameters: Map<String, String> = emptyMap()
 ) {
 
-  fun createDiscoveryRequest(): LauncherDiscoveryRequest =
-      LauncherDiscoveryRequestBuilder.request()
-          .selectors(this.selectors)
-          .filters(*this.filters.toTypedArray())
-          .configurationParameters(this.configurationParameters)
-          .build()
+    fun createDiscoveryRequest(): LauncherDiscoveryRequest =
+        LauncherDiscoveryRequestBuilder.request()
+            .selectors(this.selectors)
+            .filters(*this.filters.toTypedArray())
+            .configurationParameters(this.configurationParameters)
+            .build()
+
+    fun isIsolatedMethodRun(): Boolean {
+        return selectors.size == 1 && selectors.first() is MethodSelector
+    }
 }